requirements.txt
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
Flask==2.3.3
Flask-Cors==4.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
pandas==2.2.3
paramiko==3.5.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==1.7.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.2
schedule==1.2.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
